Missed some renaming spots
