+++ cli.js
@@ -46,7 +46,7 @@ prog
                 await changelogGenerator.call(context, client, tags);
                 logger.info('Changelog Generated as CHANGELOG.md');
             } else {
-                throw new TypeError('Error: Config options are not set in command line or in .changelog-generator-config.json. See --help to learn what you can send.');
+                throw new TypeError('Error: Config options are not set in command line or in .origami-logs-config.json. See --help to learn what you can send.');
             }
         } catch (e) {
             logger.error(e.message);
